infra: 프로덕션 인프라 임시 파일 제거 및 마이크로서비스 인프라 적용
