+++ aws-cdk/lib/env-agnostic/domain-stack.ts
@@ -7,7 +7,6 @@ import { constants } from '../../constants';
 interface LCDomainStackProps extends cdk.StackProps {
   prodALB: elbv2.ApplicationLoadBalancer;
   devALB?: elbv2.ApplicationLoadBalancer;
-  prodPrivateAlb: elbv2.ApplicationLoadBalancer;
 }
 
 export class LCDomainStack extends cdk.Stack {
@@ -16,14 +15,13 @@ export class LCDomainStack extends cdk.Stack {
 
   private readonly prodALBTarget: route53Targets.LoadBalancerTarget;
   private readonly devALBTarget: route53Targets.LoadBalancerTarget;
-  private readonly prodPrivateAlbTarget: route53Targets.LoadBalancerTarget;
 
   public hostedzone: route53.PublicHostedZone;
 
   constructor(scope: cdk.Construct, id: string, props: LCDomainStackProps) {
     super(scope, id, props);
 
-    const { prodALB, devALB, prodPrivateAlb } = props;
+    const { prodALB, devALB } = props;
     if (!devALB) {
       throw new Error('dev ALB is not defined - from LCDomainStack');
     }
@@ -32,7 +30,6 @@ export class LCDomainStack extends cdk.Stack {
     this.createPublicHostedZone();
     this.prodALBTarget = new route53Targets.LoadBalancerTarget(prodALB);
     this.devALBTarget = new route53Targets.LoadBalancerTarget(devALB);
-    this.prodPrivateAlbTarget = new route53Targets.LoadBalancerTarget(prodPrivateAlb);
 
     this.createProdRecords();
     this.createDevRecords();
@@ -85,12 +82,6 @@ export class LCDomainStack extends cdk.Stack {
       recordName: `overlay-controller.${this.PUNYCODE_DOMAIN}`,
       zone: this.hostedzone,
       target: route53.RecordTarget.fromAlias(this.prodALBTarget),
-    });
-    // mailer.크크쇼.com
-    new route53.ARecord(this, `${this.PUNYCODE_DOMAIN}_ARecord_mailer`, {
-      zone: this.hostedzone,
-      recordName: `mailer.${this.PUNYCODE_DOMAIN}`,
-      target: route53.RecordTarget.fromAlias(this.prodPrivateAlbTarget),
     });
   }
