+++ aws-cdk/lib/prod/vpc-stack.ts
@@ -9,7 +9,6 @@ const DATABASE_PORT = 3306;
 export class LCProdVpcStack extends cdk.Stack {
   public readonly vpc: ec2.Vpc;
   public readonly albSecGrp: ec2.SecurityGroup;
-  public readonly privateAlbSecGrp: ec2.SecurityGroup;
   public readonly dbSecGrp: ec2.SecurityGroup;
   public readonly apiSecGrp: ec2.SecurityGroup;
   public readonly overlaySecGrp: ec2.SecurityGroup;
@@ -28,19 +27,14 @@ export class LCProdVpcStack extends cdk.Stack {
     // * vpc
     this.vpc = this.createVpc();
     this.albSecGrp = this.createAlbSecGrp();
-    this.privateAlbSecGrp = this.createPrivateAlbSecGrp();
-    this.apiSecGrp = this.createApiSecGrp(this.albSecGrp);
-    this.overlaySecGrp = this.createOverlaySecGrp(this.albSecGrp);
-    this.overlayControllerSecGrp = this.createOverlayControllerSecGrp(this.albSecGrp);
-    this.realtimeApiSecGrp = this.createRealtimeApiSecGrp(this.albSecGrp);
-    this.mailerSecGrp = this.createMailerSecGrp(this.albSecGrp);
+    this.redisSecGrp = this.createRedisSecGrp();
+
+    this.apiSecGrp = this.createApiSecGrp();
+    this.overlaySecGrp = this.createOverlaySecGrp();
+    this.overlayControllerSecGrp = this.createOverlayControllerSecGrp();
+    this.realtimeApiSecGrp = this.createRealtimeApiSecGrp();
+    this.mailerSecGrp = this.createMailerSecGrp();
     this.inactiveBatchSecGrp = this.createInactiveBatchSecGrp();
-    this.redisSecGrp = this.createRedisSecGrp({
-      realtimeApiSecGrp: this.realtimeApiSecGrp,
-      apiSecGrp: this.apiSecGrp,
-      overlaySecGrp: this.overlaySecGrp,
-      overlayControllerSecGrp: this.overlayControllerSecGrp,
-    });
     this.dbSecGrp = this.createDbSecGrp({
       apiSecGrp: this.apiSecGrp,
       overlaySecGrp: this.overlaySecGrp,
@@ -84,65 +78,33 @@ export class LCProdVpcStack extends cdk.Stack {
     return albSecGrp;
   }
 
-  /** 프라이빗 로드밸런서(ALB) 보안 그룹 생성 */
-  private createPrivateAlbSecGrp(): ec2.SecurityGroup {
-    const privateAlbSecGrp = new ec2.SecurityGroup(
-      this,
-      `${ID_PREFIX}Private-ALB-SecGrp`,
-      {
-        vpc: this.vpc,
-        description: 'private ALB security group for project-lc',
-        allowAllOutbound: true,
-      },
-    );
-
-    privateAlbSecGrp.addIngressRule(
-      ec2.Peer.anyIpv4(),
-      ec2.Port.tcp(80),
-      'Allow 80 to all',
-    );
-    privateAlbSecGrp.addIngressRule(
-      ec2.Peer.anyIpv4(),
-      ec2.Port.tcp(443),
-      'Allow 443 to all',
-    );
-
-    return privateAlbSecGrp;
-  }
-
   /** API서버 보안그룹 생성 */
-  private createApiSecGrp(albSecGrp: ec2.SecurityGroup): ec2.SecurityGroup {
+  private createApiSecGrp(): ec2.SecurityGroup {
     const apiSecGrp = new ec2.SecurityGroup(this, `${ID_PREFIX}API-SecGrp`, {
       vpc: this.vpc,
       description: 'api security grp for project-lc',
       allowAllOutbound: true,
     });
-    apiSecGrp.addIngressRule(
-      albSecGrp,
-      ec2.Port.tcp(constants.PROD.ECS_API_PORT),
-      'allow port 3000',
-    );
+    this.allowAlbToServer(apiSecGrp, constants.PROD.ECS_API_PORT);
+    this.allowRedisToServer(apiSecGrp, constants.PROD.ECS_API_PORT);
     return apiSecGrp;
   }
 
   /** Overlay서버 보안그룹 생성 */
-  private createOverlaySecGrp(albSecGrp: ec2.SecurityGroup): ec2.SecurityGroup {
+  private createOverlaySecGrp(): ec2.SecurityGroup {
     const overlaySecGrp = new ec2.SecurityGroup(this, `${ID_PREFIX}Overlay-SecGrp`, {
       vpc: this.vpc,
       description: 'overlay security grp for project-lc',
       allowAllOutbound: true,
     });
 
-    overlaySecGrp.addIngressRule(
-      albSecGrp,
-      ec2.Port.tcp(constants.PROD.ECS_OVERLAY_PORT),
-      'allow port 3002',
-    );
+    this.allowAlbToServer(overlaySecGrp, constants.PROD.ECS_OVERLAY_PORT);
+    this.allowRedisToServer(overlaySecGrp, constants.PROD.ECS_OVERLAY_PORT);
     return overlaySecGrp;
   }
 
   /** 오버레이 컨트롤러 서버 보안 그룹 생성 */
-  private createOverlayControllerSecGrp(albSecGrp: ec2.SecurityGroup): ec2.SecurityGroup {
+  private createOverlayControllerSecGrp(): ec2.SecurityGroup {
     const overlayControllerSecGrp = new ec2.SecurityGroup(
       this,
       `${ID_PREFIX}OverlayController-SecGrp`,
@@ -153,10 +115,13 @@ export class LCProdVpcStack extends cdk.Stack {
       },
     );
 
-    overlayControllerSecGrp.addIngressRule(
-      albSecGrp,
-      ec2.Port.tcp(constants.PROD.ECS_OVERLAY_CONTROLLER_PORT),
-      'allow port 3333 to alb',
+    this.allowAlbToServer(
+      overlayControllerSecGrp,
+      constants.PROD.ECS_OVERLAY_CONTROLLER_PORT,
+    );
+    this.allowRedisToServer(
+      overlayControllerSecGrp,
+      constants.PROD.ECS_OVERLAY_CONTROLLER_PORT,
     );
     return overlayControllerSecGrp;
   }
@@ -182,22 +147,22 @@ export class LCProdVpcStack extends cdk.Stack {
     dbSecGrp.addIngressRule(
       ec2.Peer.ipv4(constants.WHILETRUE_IP_ADDRESS),
       ec2.Port.tcp(DATABASE_PORT),
-      'Allow port 3306 for outbound traffics to the whiletrue developers',
+      `Allow port ${DATABASE_PORT} for outbound traffics to the whiletrue developers`,
     );
     dbSecGrp.addIngressRule(
       apiSecGrp ?? this.apiSecGrp,
       ec2.Port.tcp(DATABASE_PORT),
-      'Allow port 3306 only to traffic from api security group',
+      `Allow port ${DATABASE_PORT} only to traffic from api security group`,
     );
     dbSecGrp.addIngressRule(
       overlaySecGrp ?? this.overlaySecGrp,
       ec2.Port.tcp(DATABASE_PORT),
-      'Allow port 3306 only to traffic from overlay security group',
+      `Allow port ${DATABASE_PORT} only to traffic from overlay security group`,
     );
     dbSecGrp.addIngressRule(
       overlayControllerSecGrp ?? this.overlayControllerSecGrp,
       ec2.Port.tcp(DATABASE_PORT),
-      'Allow port 3306 only to traffic from overlay controller security group',
+      `Allow port ${DATABASE_PORT} only to traffic from overlay controller security group`,
     );
 
     const githubActionsRunnerSecGrp = new ec2.SecurityGroup(
@@ -230,7 +195,7 @@ export class LCProdVpcStack extends cdk.Stack {
   }
 
   /** realtime API 서버 보안그룹 생성 */
-  private createRealtimeApiSecGrp(albSecGrp: ec2.SecurityGroup): ec2.SecurityGroup {
+  private createRealtimeApiSecGrp(): ec2.SecurityGroup {
     const realtimeApiSecGrp = new ec2.SecurityGroup(
       this,
       `${ID_PREFIX}RealtimeApi-SecGrp`,
@@ -241,115 +206,38 @@ export class LCProdVpcStack extends cdk.Stack {
       },
     );
 
-    realtimeApiSecGrp.addIngressRule(
-      albSecGrp,
-      ec2.Port.tcp(constants.PROD.ECS_REALTIME_API_PORT),
-      'Allow port 3001 to alb for kks realtime-api',
-    );
+    this.allowAlbToServer(realtimeApiSecGrp, constants.PROD.ECS_REALTIME_API_PORT);
+    this.allowRedisToServer(realtimeApiSecGrp, constants.PROD.ECS_REALTIME_API_PORT);
 
     return realtimeApiSecGrp;
   }
 
   /** Redis 서버 (ElstiCache) 보안그룹 생성 */
-  private createRedisSecGrp({
-    apiSecGrp,
-    overlaySecGrp,
-    overlayControllerSecGrp,
-    realtimeApiSecGrp,
-  }: {
-    apiSecGrp: ec2.SecurityGroup;
-    overlaySecGrp: ec2.SecurityGroup;
-    overlayControllerSecGrp: ec2.SecurityGroup;
-    realtimeApiSecGrp: ec2.SecurityGroup;
-  }): ec2.SecurityGroup {
+  private createRedisSecGrp(): ec2.SecurityGroup {
     const redisPort = 6379;
     const redisSecGrp = new ec2.SecurityGroup(this, `${ID_PREFIX}Redis-SecGrp`, {
       vpc: this.vpc,
       description: 'Redis security grp for project-lc',
-      allowAllOutbound: false,
+      allowAllOutbound: true,
     });
 
-    // * Allow API Server
-    this.allowRedisToServer({
-      redisSecGrp,
-      serverName: 'api',
-      secGrp: apiSecGrp,
-      redisPort,
-      serverPort: constants.PROD.ECS_API_PORT,
-    });
-    // * Allow Realtime API Server
-    this.allowRedisToServer({
-      redisSecGrp,
-      serverName: 'realtime-api',
-      secGrp: realtimeApiSecGrp,
-      redisPort,
-      serverPort: constants.PROD.ECS_REALTIME_API_PORT,
-    });
-    // * Allow Overlay Server
-    this.allowRedisToServer({
-      redisSecGrp,
-      serverName: 'overlay',
-      secGrp: overlaySecGrp,
-      redisPort,
-      serverPort: constants.PROD.ECS_OVERLAY_PORT,
-    });
-    // * Allow Overlay-controller Server
-    this.allowRedisToServer({
-      redisSecGrp,
-      serverName: 'overlay-controller',
-      secGrp: overlayControllerSecGrp,
-      redisPort,
-      serverPort: constants.PROD.ECS_OVERLAY_CONTROLLER_PORT,
-    });
+    redisSecGrp.addIngressRule(
+      ec2.Peer.anyIpv4(),
+      ec2.Port.tcp(redisPort),
+      `allow port ${redisPort} to anyIpv4`,
+    );
 
     return redisSecGrp;
   }
 
-  private allowRedisToServer = ({
-    serverName,
-    redisSecGrp,
-    secGrp,
-    redisPort = 6379,
-    serverPort,
-  }: {
-    serverName: string;
-    redisSecGrp: ec2.SecurityGroup;
-    secGrp: ec2.SecurityGroup;
-    serverPort: number;
-    redisPort?: number;
-  }): void => {
-    // * API server
-    redisSecGrp.addIngressRule(
-      secGrp,
-      ec2.Port.tcp(redisPort),
-      `allow port ${redisPort} to ${serverName} server`,
-    );
-    redisSecGrp.addEgressRule(
-      secGrp,
-      ec2.Port.tcp(redisPort),
-      `allow port ${redisPort} to ${serverName} server`,
-    );
-    secGrp.addIngressRule(
-      redisSecGrp,
-      ec2.Port.tcp(serverPort),
-      `allow port ${serverPort} to redis cluster`,
-    );
-  };
-
   /** Mailer 서버 보안그룹 생성 */
-  private createMailerSecGrp(albSecGrp: ec2.SecurityGroup): ec2.SecurityGroup {
+  private createMailerSecGrp(): ec2.SecurityGroup {
     const mailerSecGrp = new ec2.SecurityGroup(this, `${ID_PREFIX}Mailer-SecGrp`, {
       vpc: this.vpc,
       description: 'mailer sec grp for project-lc (private)',
       allowAllOutbound: true,
     });
-
-    mailerSecGrp.addIngressRule(
-      albSecGrp,
-      ec2.Port.tcp(constants.PROD.ECS_MAILER_PORT),
-      'Allow port 3003 to public alb',
-    );
-
+    this.allowRedisToServer(mailerSecGrp, constants.PROD.ECS_MAILER_PORT);
     return mailerSecGrp;
   }
 
@@ -365,5 +253,35 @@ export class LCProdVpcStack extends cdk.Stack {
       },
     );
     return inactiveBatchSecGrp;
+  }
+
+  // ******************************
+  // * 보안 그룹 적용 헬퍼 함수
+  // ******************************
+
+  /** 서버보안그룹에 Redis로부터의 요청을 허용하는 메서드로, 레디스 보안그룹 생성 이후 사용해야 함. */
+  private allowRedisToServer(secGrp: ec2.SecurityGroup, serverPort: number): void {
+    if (!this.redisSecGrp)
+      throw Error(
+        '서버보안그룹에 Redis로부터의 요청을 허용하는 메서드로, 레디스 보안그룹 생성 이후 사용해야 함.',
+      );
+    secGrp.addIngressRule(
+      this.redisSecGrp,
+      ec2.Port.tcp(serverPort),
+      `allow port ${serverPort} to redis cluster`,
+    );
+  }
+
+  /** 서버보안그룹에 ALB로부터의 요청을 허용하는 메서드로, 로드밸런서 보안그룹 생성 이후 사용해야 함. */
+  private allowAlbToServer(secGrp: ec2.SecurityGroup, serverPort: number): void {
+    if (!this.albSecGrp)
+      throw Error(
+        '서버보안그룹에 ALB로부터의 요청을 허용하는 메서드로, 로드밸런서 보안그룹 생성 이후 사용해야 함.',
+      );
+    secGrp.addIngressRule(
+      this.albSecGrp,
+      ec2.Port.tcp(serverPort),
+      `allow port ${serverPort} to ALB`,
+    );
   }
 }
